SSCS-10116 Moving ReferenceData to service
