+++ src/main/java/uk/gov/hmcts/reform/sscs/service/ReferenceData.java
@@ -1,10 +1,8 @@
-package uk.gov.hmcts.reform.sscs.model;
+package uk.gov.hmcts.reform.sscs.service;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.sscs.service.HearingDurationsService;
-import uk.gov.hmcts.reform.sscs.service.SessionCategoryMapService;
 
 @Data
 @AllArgsConstructor
